replace size_t to uint32_t, thats descreased register usage

+++ ScalarFunction.cpp
@@ -3,16 +3,16 @@
 #include <boost/thread.hpp>
 #include <boost/shared_ptr.hpp>
 
-void ScalarFunction::makeConvex( const size_t& dimX, const size_t& numberOfPoints )
+void ScalarFunction::makeConvex( const uint32_t& dimX, const uint32_t& numberOfPoints )
 {
 	if( dimX == 0 )
 		return;
 	
 	FP dFi = PI / ( numberOfPoints - 1 );
 
-	size_t n = dimX + 1; // space dimension
+	uint32_t n = dimX + 1; // space dimension
 
-	size_t numberOfHyperplanes = pow( numberOfPoints, n - 1 );
+	uint32_t numberOfHyperplanes = pow( numberOfPoints, n - 1 );
 
 	// first x0.. x(n - 2) elements are independent vars. in 2D it will be x, contains normal
 	// x(n - 1) element dependent var. . in 2D it will be y
@@ -21,12 +21,12 @@ void ScalarFunction::makeConvex( const size_t& dimX, const size_t& numberOfPoint
 
 	FPVector fi( dimX, 0.0 );
 
-	for( size_t i = 0; i < numberOfHyperplanes; i++ )
+	for( uint32_t i = 0; i < numberOfHyperplanes; i++ )
 	{
-		for( size_t j = 0; j < n; j++ )
+		for( uint32_t j = 0; j < n; j++ )
 		{
 			hyperplanes[ i ][ j ] = 1.0;
-			for( size_t k = 0; k < j; k++ )
+			for( uint32_t k = 0; k < j; k++ )
 				hyperplanes[ i ][ j ] *= sin( fi[ k ] );
 
 			if( j != n - 1 )
@@ -35,7 +35,7 @@ void ScalarFunction::makeConvex( const size_t& dimX, const size_t& numberOfPoint
 
 		// not good enough
 		bool shift = true;
-		for( size_t k = 0; ( k < dimX ) && shift; k++ )
+		for( uint32_t k = 0; ( k < dimX ) && shift; k++ )
 		{
 			if( fabs( fi[ k ] - PI ) <= EPSILON )
 			{
@@ -59,7 +59,7 @@ void ScalarFunction::makeConvex( const size_t& dimX, const size_t& numberOfPoint
 			FP d = 0.0;
 
 			// dot product of point and normal is distance
-			for( size_t j = 0; j < dimX; j++ )
+			for( uint32_t j = 0; j < dimX; j++ )
 				d += iter->first[ j ] * hyperplanes[ i ][ j ];
 			d += iter->second * hyperplanes[ i ][ n - 1 ];
 
@@ -71,13 +71,13 @@ void ScalarFunction::makeConvex( const size_t& dimX, const size_t& numberOfPoint
 	for( iterator iter = begin(); iter != end(); ++iter )
 	{
 		FP funcVal = iter->second;
-		for( size_t i = 0; i < numberOfHyperplanes; i++ )
+		for( uint32_t i = 0; i < numberOfHyperplanes; i++ )
 		{
 			FP val = 0.0;
 			// xi - iter->first
 			// Ni - hyperplane normal
 			// val = x(n - 1) = ( -N0*x0 - N1*x1 - ... - N(n - 2)*x(n - 2) + xn ) / N(n - 1)
-			for( size_t j = 0; j < dimX; j++ )
+			for( uint32_t j = 0; j < dimX; j++ )
 				val -= iter->first[ j ] * hyperplanes[ i ][ j ];
 			val += hyperplanes[ i ][ n ];
 			val /= hyperplanes[ i ][ n - 1 ] + EPSILON;
@@ -97,17 +97,17 @@ void ScalarFunction::makeConvex( const size_t& dimX, const size_t& numberOfPoint
 
 //
 void thread1( std::vector< FPVector >& hyperplanes, std::vector< FPVector >& fi, 
-			const size_t& startIndex, const size_t& taskSize, const size_t& n, 
+			const uint32_t& startIndex, const uint32_t& taskSize, const uint32_t& n, 
 			ScalarFunction* func )
 {
-	size_t dimX = n - 1;
+	uint32_t dimX = n - 1;
 
-	for( size_t i = startIndex; i < startIndex + taskSize; i++ )
+	for( uint32_t i = startIndex; i < startIndex + taskSize; i++ )
 	{
-		for( size_t j = 0; j < n; j++ )
+		for( uint32_t j = 0; j < n; j++ )
 		{
 			hyperplanes[ i ][ j ] = 1.0;
-			for( size_t k = 0; k < j; k++ )
+			for( uint32_t k = 0; k < j; k++ )
 				hyperplanes[ i ][ j ] *= sin( fi[ i ][ k ] );
 
 			if( j != n - 1 )
@@ -121,7 +121,7 @@ void thread1( std::vector< FPVector >& hyperplanes, std::vector< FPVector >& fi,
 			FP d = 0.0;
 
 			// dot product of point and normal is distance
-			for( size_t j = 0; j < dimX; j++ )
+			for( uint32_t j = 0; j < dimX; j++ )
 				d += iter->first[ j ] * hyperplanes[ i ][ j ];
 			d += iter->second * hyperplanes[ i ][ n - 1 ];
 
@@ -133,23 +133,23 @@ void thread1( std::vector< FPVector >& hyperplanes, std::vector< FPVector >& fi,
 
 
 void thread2( std::vector< FPVector >& hyperplanes, std::vector< const FPVector* >& points, 
-			const size_t& startIndex, const size_t& taskSize, const size_t& n, 
+			const uint32_t& startIndex, const uint32_t& taskSize, const uint32_t& n, 
 			ScalarFunction* func )
 {
-	size_t dimX = n - 1;
+	uint32_t dimX = n - 1;
 
-	for( size_t k = startIndex; k < startIndex + taskSize; k++ )
+	for( uint32_t k = startIndex; k < startIndex + taskSize; k++ )
 	{
 		FP funcVal = func->at( *points[ k ] );
 		FP ret = funcVal;
 
-		for( size_t i = 0; i < hyperplanes.size(); i++ )
+		for( uint32_t i = 0; i < hyperplanes.size(); i++ )
 		{
 			FP val = 0.0;
 			// xi - iter->first
 			// Ni - hyperplane normal
 			// val = x(n - 1) = ( -N0*x0 - N1*x1 - ... - N(n - 2)*x(n - 2) + xn ) / N(n - 1)
-			for( size_t j = 0; j < dimX; j++ )
+			for( uint32_t j = 0; j < dimX; j++ )
 				val -= ( *points[ k ] )[ j ] * hyperplanes[ i ][ j ];
 			val += hyperplanes[ i ][ n ];
 			val /= hyperplanes[ i ][ n - 1 ] + EPSILON;
@@ -178,16 +178,16 @@ void JoinThread( boost::thread* thread )
 
 
 //
-void ScalarFunction::makeConvexMultiThread( const size_t& dimX, const size_t& numberOfPoints, const size_t& jobs )
+void ScalarFunction::makeConvexMultiThread( const uint32_t& dimX, const uint32_t& numberOfPoints, const uint32_t& jobs )
 {
 	if( dimX == 0 )
 		return;
 	
 	FP dFi = PI / ( numberOfPoints - 1 );
 
-	size_t n = dimX + 1; // space dimension
+	uint32_t n = dimX + 1; // space dimension
 
-	size_t numberOfHyperplanes = pow( numberOfPoints, n - 1 );
+	uint32_t numberOfHyperplanes = pow( numberOfPoints, n - 1 );
 
 	FPVector normal( n );
 
@@ -198,11 +198,11 @@ void ScalarFunction::makeConvexMultiThread( const size_t& dimX, const size_t& nu
 
 	// prepare Fi for all hyperplanes
 	std::vector< FPVector > fi( numberOfHyperplanes, FPVector( dimX, 0.0 ) );
-	for( size_t i = 1; i < numberOfHyperplanes; i++ )
+	for( uint32_t i = 1; i < numberOfHyperplanes; i++ )
 	{
 		bool shift = true;
 		fi[ i ] = fi[ i - 1 ];
-		for( size_t k = 0; ( k < dimX ) && shift; k++ )
+		for( uint32_t k = 0; ( k < dimX ) && shift; k++ )
 		{
 			if( fabs( fi[ i - 1 ][ k ] - PI ) <= EPSILON )
 			{
@@ -223,13 +223,13 @@ void ScalarFunction::makeConvexMultiThread( const size_t& dimX, const size_t& nu
 	{
 		std::vector< boost::shared_ptr< boost::thread > > threads( jobs );
 
-		size_t tasksNumber = numberOfHyperplanes;
+		uint32_t tasksNumber = numberOfHyperplanes;
 
-		size_t tasksPerThread = tasksNumber / jobs;
+		uint32_t tasksPerThread = tasksNumber / jobs;
 
-		for( size_t i = 0; i < jobs; i++ )
+		for( uint32_t i = 0; i < jobs; i++ )
 		{
-			size_t tasksNumberForThread = i == jobs - 1 ? tasksNumber - i * tasksPerThread : tasksPerThread;
+			uint32_t tasksNumberForThread = i == jobs - 1 ? tasksNumber - i * tasksPerThread : tasksPerThread;
 			threads[ i ] = boost::shared_ptr< boost::thread >( new boost::thread( thread1, boost::ref( hyperplanes ), boost::ref( fi ), i * tasksPerThread, 
 																tasksNumberForThread, n, this ),
 																&JoinThread );
@@ -238,20 +238,20 @@ void ScalarFunction::makeConvexMultiThread( const size_t& dimX, const size_t& nu
 
 	std::vector< const FPVector* > points( size() );
 
-	size_t i = 0;
+	uint32_t i = 0;
 	for( iterator iter = begin(); iter != end(); ++iter, i++ )
 		points[ i ] = &iter->first;
 
 	{
 		std::vector< boost::shared_ptr< boost::thread > > threads( jobs );
 
-		size_t tasksNumber = size();
+		uint32_t tasksNumber = size();
 
-		size_t tasksPerThread = tasksNumber / jobs;
+		uint32_t tasksPerThread = tasksNumber / jobs;
 
-		for( size_t i = 0; i < jobs; i++ )
+		for( uint32_t i = 0; i < jobs; i++ )
 		{
-			size_t tasksNumberForThread = i == jobs - 1 ? tasksNumber - i * tasksPerThread : tasksPerThread;
+			uint32_t tasksNumberForThread = i == jobs - 1 ? tasksNumber - i * tasksPerThread : tasksPerThread;
 			threads[ i ] = boost::shared_ptr< boost::thread >( new boost::thread( thread2, boost::ref( hyperplanes ), boost::ref( points ), i * tasksPerThread, 
 																tasksNumberForThread, n, this ),
 																&JoinThread );
@@ -263,7 +263,7 @@ void ScalarFunction::makeConvexMultiThread( const size_t& dimX, const size_t& nu
 #ifdef GPU
 #include "gpu.hpp"
 
-void ScalarFunction::makeConvexGPU( const size_t& dimX, const size_t& numberOfPoints )
+void ScalarFunction::makeConvexGPU( const uint32_t& dimX, const uint32_t& numberOfPoints )
 {
 	gpu::makeConvex( *this, dimX, numberOfPoints );
 }
